+++ gluten-core/src/main/scala/org/apache/gluten/extension/columnar/ColumnarRuleApplier.scala
@@ -1,0 +1,7 @@
+package org.apache.gluten.extension.columnar
+
+import org.apache.spark.sql.execution.SparkPlan
+
+trait ColumnarRuleApplier {
+  def apply(plan: SparkPlan, outputsColumnar: Boolean): SparkPlan
+}
