+++ gluten-core/src/main/scala/org/apache/gluten/planner/GlutenOptimization.scala
@@ -20,7 +20,6 @@ import org.apache.gluten.planner.cost.GlutenCostModel
 import org.apache.gluten.planner.metadata.GlutenMetadataModel
 import org.apache.gluten.planner.plan.GlutenPlanModel
 import org.apache.gluten.planner.property.GlutenPropertyModel
-import org.apache.gluten.planner.rule.GlutenRules
 import org.apache.gluten.ras.{Optimization, RasExplain}
 import org.apache.gluten.ras.rule.RasRule
 
@@ -29,16 +28,6 @@ import org.apache.spark.sql.execution.SparkPlan
 object GlutenOptimization {
   private object GlutenExplain extends RasExplain[SparkPlan] {
     override def describeNode(node: SparkPlan): String = node.nodeName
-  }
-
-  def apply(): Optimization[SparkPlan] = {
-    Optimization[SparkPlan](
-      GlutenPlanModel(),
-      GlutenCostModel(),
-      GlutenMetadataModel(),
-      GlutenPropertyModel(),
-      GlutenExplain,
-      RasRule.Factory.reuse(GlutenRules()))
   }
 
   def apply(rules: Seq[RasRule[SparkPlan]]): Optimization[SparkPlan] = {
