+++ gluten-ut/spark32/src/test/scala/org/apache/spark/sql/execution/FallbackStrategiesSuite.scala
@@ -17,8 +17,9 @@
 package org.apache.spark.sql.execution
 
 import org.apache.gluten.execution.BasicScanExecTransformer
-import org.apache.gluten.extension.{ColumnarOverrideRules, GlutenPlan}
+import org.apache.gluten.extension.GlutenPlan
 import org.apache.gluten.extension.columnar.InsertTransitions
+import org.apache.gluten.extension.columnar.heuristic.HeuristicApplier
 import org.apache.gluten.utils.QueryPlanSelector
 
 import org.apache.spark.rdd.RDD
@@ -31,7 +32,7 @@ class FallbackStrategiesSuite extends GlutenSQLTestsTrait {
   testGluten("Fall back the whole query if one unsupported") {
     withSQLConf(("spark.gluten.sql.columnar.query.fallback.threshold", "1")) {
       val originalPlan = UnaryOp2(UnaryOp1(UnaryOp2(UnaryOp1(LeafOp()))))
-      val rule = ColumnarOverrideRules(spark).withTransformRules(
+      val rule = new HeuristicApplier(spark).withTransformRules(
         List(
           _ =>
             _ => {
@@ -47,7 +48,7 @@ class FallbackStrategiesSuite extends GlutenSQLTestsTrait {
   testGluten("Fall back the whole plan if meeting the configured threshold") {
     withSQLConf(("spark.gluten.sql.columnar.wholeStage.fallback.threshold", "1")) {
       val originalPlan = UnaryOp2(UnaryOp1(UnaryOp2(UnaryOp1(LeafOp()))))
-      val rule = ColumnarOverrideRules(spark)
+      val rule = new HeuristicApplier(spark)
         .enableAdaptiveContext()
         .withTransformRules(
           List(
@@ -65,7 +66,7 @@ class FallbackStrategiesSuite extends GlutenSQLTestsTrait {
   testGluten("Don't fall back the whole plan if NOT meeting the configured threshold") {
     withSQLConf(("spark.gluten.sql.columnar.wholeStage.fallback.threshold", "4")) {
       val originalPlan = UnaryOp2(UnaryOp1(UnaryOp2(UnaryOp1(LeafOp()))))
-      val rule = ColumnarOverrideRules(spark)
+      val rule = new HeuristicApplier(spark)
         .enableAdaptiveContext()
         .withTransformRules(
           List(
@@ -85,7 +86,7 @@ class FallbackStrategiesSuite extends GlutenSQLTestsTrait {
       " transformable)") {
     withSQLConf(("spark.gluten.sql.columnar.wholeStage.fallback.threshold", "2")) {
       val originalPlan = UnaryOp2(UnaryOp1(UnaryOp2(UnaryOp1(LeafOp()))))
-      val rule = ColumnarOverrideRules(spark)
+      val rule = new HeuristicApplier(spark)
         .enableAdaptiveContext()
         .withTransformRules(
           List(
@@ -105,7 +106,7 @@ class FallbackStrategiesSuite extends GlutenSQLTestsTrait {
       "leaf node is transformable)") {
     withSQLConf(("spark.gluten.sql.columnar.wholeStage.fallback.threshold", "3")) {
       val originalPlan = UnaryOp2(UnaryOp1(UnaryOp2(UnaryOp1(LeafOp()))))
-      val rule = ColumnarOverrideRules(spark)
+      val rule = new HeuristicApplier(spark)
         .enableAdaptiveContext()
         .withTransformRules(
           List(
